requirements.txt
python-dotenv
pydantic
pandas
MetaTrader5
python-telegram-bot